+++ src/Resources/MCPServer.Resource.Logs.pas
@@ -302,7 +302,10 @@ function TLogsByLevelTemplate.Complete(const ArgumentName, Value: string;
   const Context: TArray<TPair<string, string>>): TMCPCompletion;
 begin
   if ArgumentName <> TEMPLATE_VARIABLE_LEVEL then
-    Exit(TMCPCompletion.Create(nil));
+    begin
+      Result := TMCPCompletion.Create(nil);
+      Exit;
+    end;
 
   var Levels := TStringList.Create;
   try